UPDATE: add multiprocessing and simplify code
